Added simple Student_Life and On_Campus pages
Created a simple student life page with On campus and Off Campus links on it. Really bare bones at the moment could be expanded on.

Also created the On Campus page. Not yet finished. Need to expand on clubs and events. Society section mostly done.

Added .bak files to .gitignore since the editor i'm using creates them but they arent needed for the site.